git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/BAM-COW@24460 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/automaton/AutomatonStatistics.java
@@ -61,7 +61,7 @@ class AutomatonStatistics extends AbstractStatistics {
       put(out, 0, trans.totalStrengthenTime);
     }
 
-    int stateBranchings = trans.automatonSuccessors.getValueCount()
+    long stateBranchings = trans.automatonSuccessors.getValueCount()
         - trans.automatonSuccessors.getTimesWithValue(0)
         - trans.automatonSuccessors.getTimesWithValue(1);
     put(out, 0, "Automaton transfers with branching", stateBranchings);
